% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{adjust_negative_predicted_features}
\alias{adjust_negative_predicted_features}
\title{Function to adjust any negative predicted features}
\usage{
adjust_negative_predicted_features(
  df,
  min_lowest_feature_val = 0.006,
  grouping_variables = c("country", "year")
)
}
\arguments{
\item{df}{dataframe with decile data}

\item{min_lowest_feature_val}{Lowest possible value for lowest feature (e.g. d1).}

\item{grouping_variables}{variables to group by}
}
\description{
Sometimes PC model can generate negative values for some features. E.g. South Africa where GINI is much too high.
Here, the function adjusts negative predicted features.
}
