% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{aggregate_country_deciles_to_regions}
\alias{aggregate_country_deciles_to_regions}
\title{Function to aggregate country deciles to regions}
\usage{
aggregate_country_deciles_to_regions(
  df,
  value_col = "Income..net.",
  grouping_variables = c("GCAM_region_ID", "year")
)
}
\arguments{
\item{df}{dataframe with decile data}

\item{grouping_variables}{variables to group by. This should point to the new regional mapping and any other dimension that you want to aggregate to.}
}
\description{
Function to aggregate country deciles to regions
}
