% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{compute_components_data}
\alias{compute_components_data}
\title{Function to calculate coefficients from raw data}
\usage{
compute_components_data(
  df,
  center_and_scaler_data = pc_center_sd,
  pc_loadings = pc_loading_matrix,
  category_col = "Category",
  value_col = "value",
  SCALE = TRUE,
  number_of_features = 10,
  grouping_variables = c("country", "year")
)
}
\arguments{
\item{df}{dataframe with decile data}

\item{center_and_scaler_data}{The centers and SD for the data. This is pre-saved. But can be generated using `get_sd_center()` above.}

\item{category_col}{column with categorical definitions for deciles}

\item{value_col}{Name of values to be output.}

\item{SCALE}{Boolean which defines if the PC analysis is to be scaled or not.}

\item{grouping_variables}{variables to group by}
}
\description{
Function to calculate coefficients from raw data
}
