% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{compute_deciles_lognormal}
\alias{compute_deciles_lognormal}
\title{Function to compute deciles using lognormal distribution}
\usage{
compute_deciles_lognormal(df, print_progress = TRUE)
}
\arguments{
\item{df}{Dataset that should contain cols- `gdp_ppp_pc_usd2011` which is income in thous USD. `gini` which is the GINI coeff
`year` year of observation, `sce` which is the scenario name}

\item{print_progress}{Boolean which specifies if progress should be printed. A txt file will be generated.}
}
\description{
Function to compute deciles using lognormal distribution
}
