% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{compute_gini_deciles}
\alias{compute_gini_deciles}
\title{Function to calculate GINI coefficient from deciles.}
\usage{
compute_gini_deciles(
  df,
  inc_col = "shares",
  grouping_variables = c("region", "year", "sce", "model", "method")
)
}
\arguments{
\item{df}{dataframe with decile data}

\item{inc_col}{column with income shares. Make sure these are normalized that is they range from 0-1}

\item{grouping_variables}{variables to group by}
}
\description{
Function to calculate GINI coefficient from deciles.
}
