% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{compute_lognormal_country}
\alias{compute_lognormal_country}
\title{Function to compute lognormal distribution for individual countries. This is passed for parallelization.}
\usage{
compute_lognormal_country(df)
}
\arguments{
\item{df}{Dataset that should contain cols- `gdp_ppp_pc_usd2011` which is income in thous USD. `gini` which is the GINI coeff
`year` year of observation, `sce` which is the scenario name}
}
\description{
Function to compute lognormal distribution for individual countries. This is passed for parallelization.
}
