% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{compute_lognormal_dist}
\alias{compute_lognormal_dist}
\title{Function to compute lognormal density distribution for arbritary params}
\usage{
compute_lognormal_dist(
  mean_income,
  gini,
  max_income = mean_income * 10,
  len_sample = 2000
)
}
\arguments{
\item{mean_income}{Average income level in thous USD}

\item{gini}{GINI coeff ranging from 0 to 1}

\item{max_income}{cutoff for max income. Set to 10 times the mean by default.}

\item{len_sample}{The number of samples to draw for lognormal. Default is set to 2000.}
}
\description{
Function to compute lognormal density distribution for arbritary params
}
