% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{compute_palma_ratio}
\alias{compute_palma_ratio}
\title{Function to calculate palma ratio from deciles}
\usage{
compute_palma_ratio(
  df,
  group_cols = c("country", "year"),
  category_col = "Category",
  income_col = "pred_shares"
)
}
\arguments{
\item{df}{dataframe with decile data}

\item{group_cols}{variables to group by}

\item{category_col}{column with categorical definitions for deciles}

\item{inc_col}{column with income shares. Make sure these are normalized that is they range from 0-1}
}
\description{
Function to calculate palma ratio from deciles
}
