% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{compute_PC_model_components}
\alias{compute_PC_model_components}
\title{Function to calculate PC coefficients for future years}
\usage{
compute_PC_model_components(
  df,
  grouping_variables = c("country", "year"),
  value_col = "Income..net."
)
}
\arguments{
\item{df}{dataframe with decile data}

\item{grouping_variables}{variables to group by}

\item{value_col}{Name of values to be output.}

\item{inc_col}{column with income shares. Make sure these are normalized that is they range from 0-1}

\item{category_col}{column with categorical definitions for deciles}
}
\description{
Function to calculate PC coefficients for future years
}
