% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{create_cores}
\alias{create_cores}
\title{parallelization function to generate multiple cores from the CPU.}
\usage{
create_cores(num_cores = NULL)
}
\description{
Note that we think this works only for Windows.To make it work for MAC, just change the `makeCluster()` type argument
in the function to FORK.
}
