% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{erfinv}
\alias{erfinv}
\title{Create inverted error function. Will need this for lognormal calculations below.}
\usage{
erfinv(x)
}
\arguments{
\item{x}{simple vector}
}
\description{
Create inverted error function. Will need this for lognormal calculations below.
}
