% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{get_centers}
\alias{get_centers}
\title{Helper function to calculate centers for the PC model.}
\usage{
get_centers(df, category_col = "category", value_col = "value")
}
\arguments{
\item{df}{dataframe with decile data}

\item{category_col}{column with categorical definitions for deciles}

\item{value_col}{Name of values to be output.}
}
\description{
Helper function to calculate centers for the PC model.
}
