% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{get_PCA_loadings}
\alias{get_PCA_loadings}
\title{Helper function to get PC loadings or components from raw data}
\usage{
get_PCA_loadings(
  df,
  category_col = "category",
  value_col = "value",
  SCALE = TRUE,
  number_of_features = 10,
  grouping_variables = c("country", "year")
)
}
\arguments{
\item{df}{dataframe with decile data}

\item{category_col}{column with categorical definitions for deciles}

\item{value_col}{Name of values to be output.}

\item{SCALE}{Boolean which defines if the PC analysis is to be scaled or not.}

\item{number_of_features}{The number of features or dimensions in data. By default is set to 10 to calculate loadings/components
for deciles.}

\item{grouping_variables}{variables to group by}
}
\description{
Helper function to get PC loadings or components from raw data
}
