% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{get_sd_center}
\alias{get_sd_center}
\title{Helper function to calculate centers and standard deviation for the PC model.}
\usage{
get_sd_center(df, category_col = "category", value_col = "value", SCALE = TRUE)
}
\arguments{
\item{df}{dataframe with decile data}

\item{category_col}{column with categorical definitions for deciles}

\item{value_col}{Name of values to be output.}

\item{SCALE}{Boolean which defines if the PC analysis is to be scaled or not.}
}
\description{
Helper function to calculate centers and standard deviation for the PC model.
}
