% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ineq.R
\name{PC_model}
\alias{PC_model}
\title{Function to calculate deciles using PC model}
\usage{
PC_model(
  df,
  grouping_variables = c("country", "year"),
  value_col = "Income..net.",
  id_col = c("country")
)
}
\arguments{
\item{df}{dataframe with all IV data. These are output by `compile_IV_data()`.}

\item{grouping_variables}{variables to group by}

\item{value_col}{Name of values to be output.}

\item{id_col}{column to split by for parallelization}
}
\description{
Function to calculate deciles using PC model
}
